// src/enums/buildings.enum.ts
enum BuildingsEnum {
    ORE_MINE = 'ORE_MINE',
    REFINER = 'REFINER',
    LIVING_MODULE = 'LIVING_MODULE',
    STORAGE = 'STORAGE',
    WORKERS_MODULE = 'WORKERS_MODULE',
    ASSEMBLY_MODULE = 'ASSEMBLY_MODULE',
    FARM_MODULE = 'FARM_MODULE',
    TRADING_MODULE = 'TRADING_MODULE',
    TELEPORT_MODULE = 'TELEPORT_MODULE',
    EXPEDITIONS_MODULE = 'EXPEDITIONS_MODULE',
    ENERGY_MODULE = 'ENERGY_MODULE',
    MAINTAIN_MODULE = 'MAINTAIN_MODULE',
    THERMOREGULATION_MODULE = 'THERMOREGULATION_MODULE',
    RADIATION_CLEANER_MODULE = 'RADIATION_CLEANER_MODULE',
    TOXICITY_CLEANER_MODULE = 'TOXICITY_CLEANER_MODULE',
    CENTRAL_MODULE = 'CENTRAL_MODULE',
    COMMUNICATION_MODULE = 'COMMUNICATION_MODULE',
    WATER_FLOAT_MODULE = 'WATER_FLOAT_MODULE',
    WATER_CLEANING_MODULE = 'WATER_CLEANING_MODULE',
    LOW_ORBIT_ENGINE_MODULE = 'LOW_ORBIT_ENGINE_MODULE'
}

export default BuildingsEnum